+++ core/plugins/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./tools.js";
 export * from "./registry/index.js";
 export * from "./datatypes.js";
+export * from "./registry/guards.js";
