+++ core/plugins/src/tools.ts
@@ -1,7 +1,7 @@
 import type { AutomergeUrl, DocHandle, Repo } from "@automerge/automerge-repo";
 import type { LoadedPlugin, PluginDescription } from "./registry/index.js";
 import { getType, type HasPatchworkMetadata } from "@patchwork/filesystem";
-import { getPlugins, sortPlugins } from "./registry/index.js";
+import { getPluginRegistry } from "./registry/index.js";
 
 import type { initializeAutomergeRepoKeyhive } from "@automerge/automerge-repo-keyhive";
 type AutomergeRepoKeyhive = Awaited<
@@ -37,12 +37,14 @@ export type Tool = LoadedPlugin<ToolDescription, ToolImplementation>;
 export type LegacyEditorProps = { docUrl: AutomergeUrl };
 
 export function getSupportedToolsForType(type: string): Tool[] {
-  const plugins = getPlugins("patchwork:tool", (desc) => {
-    return (
-      (desc as ToolDescription).supportedDataTypes.includes(type) ||
-      (desc as ToolDescription).supportedDataTypes.includes("*")
-    );
-  });
+  const plugins = getPluginRegistry<ToolDescription>("patchwork:tool").all(
+    (desc) => {
+      return (
+        desc.supportedDataTypes.includes(type) ||
+        desc.supportedDataTypes.includes("*")
+      );
+    }
+  );
 
   return plugins as Tool[];
 }
@@ -62,3 +64,51 @@ export function getFallbackTool(doc: HasPatchworkMetadata) {
     type
   )?.[0];
 }
+
+const sortPlugins = <
+  T extends LoadedPlugin<D, I>,
+  D extends PluginDescription,
+  I,
+>(
+  plugins: T[],
+  matchField: keyof D,
+  matchValue: string,
+  sortField?: keyof D
+): T[] => {
+  return [...plugins].sort((a, b) => {
+    const aValue = a[matchField];
+    const bValue = b[matchField];
+
+    // Convert string values to arrays for consistent comparison
+    const aArray = Array.isArray(aValue)
+      ? (aValue as string[])
+      : [aValue as string];
+    const bArray = Array.isArray(bValue)
+      ? (bValue as string[])
+      : [bValue as string];
+
+    const aHasWildcard = aArray.includes("*");
+    const bHasWildcard = bArray.includes("*");
+    const aHasMatch = aArray.includes(matchValue);
+    const bHasMatch = bArray.includes(matchValue);
+
+    // Specific matches come first
+    if (aHasMatch && !bHasMatch) return -1;
+    if (!aHasMatch && bHasMatch) return 1;
+
+    // Then wildcard matches come last
+    if (aHasWildcard && !bHasWildcard) return 1;
+    if (!aHasWildcard && bHasWildcard) return -1;
+
+    // If both are wildcards or both are specific matches, sort by the optional sort field
+    if (sortField) {
+      const aSort = a[sortField];
+      const bSort = b[sortField];
+      if (typeof aSort === "string" && typeof bSort === "string") {
+        return aSort.localeCompare(bSort);
+      }
+    }
+
+    return 0;
+  });
+};
